Update sdk snapshot to remove conscrypt and i18n stubs
This change removes the definitions for the following libraries:
* core.current.stubs
* core-current-stubs-system-modules
* core.module_lib.stubs
* core-module-lib-stubs-system-modules
* legacy.core.platform.api.stubs
* legacy-core-platform-api-stubs-system-modules
* stable.core.platform.api.stubs
* stable-core-platform-api-stubs-system-modules

It purposely does not remove the jars for the above modules as they
could cause conflicts if the prebuilts are updated between uploading
and submission.

And exports the following previously private members:
* core-lambda-stubs-for-system-modules
* core-generated-annotation-stubs

No additional files are needed for those as they are already present in
the snapshot.

Bug: 193311122
Test: m droid
      m SOONG_CONFIG_art_module_source_build=false nothing
Change-Id: I1eae207e79f53a9d1867e70ef7282a14077f0378